move unique static pages components to the static-pages package.
