+++ packages/component-static-pages/client/components/CalloutTextBox.js
