+++ packages/component-static-pages/client/components/List.js
