+++ packages/component-static-pages/client/components/RefTable.md
