+++ packages/component-static-pages/client/components/StaffCard.js
