+++ packages/component-static-pages/client/components/StaffCard.md
