+++ packages/component-static-pages/client/pages/ContactUs/ContactELife.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { H1 } from '@pubsweet/ui'
 
 import Paragraph from 'ui/atoms/Paragraph'
-import List from 'ui/atoms/List'
+import List from '../../components/List'
 
 const ContactELife = props => (
   <React.Fragment>
